feat: 검색 쿼리에 대한 테스트가 test containers에 띄워진 mysql db를 이용하도록 수정
